% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_percent.R
\name{as_percent}
\alias{as_percent}
\title{As Percent}
\usage{
as_percent(df = frequencies, var = result)
}
\arguments{
\item{df}{The dataframe with the number column. Defaults to the name frequencies}

\item{var}{The name of the number column. Defaults to the name result}
}
\value{
values will go from decimal to percent format like such: 0.03 -> 3%
}
\description{
This takes a column of numbers in decimal format and turns it into percent format.
}
\examples{
frequencies <- c(.30,.47,.25, .65, .42,.57,.74,.19,.54,.50,.46,.28,.45) \%>\%
as.data.frame() \%>\%
 rename(
   result = "."
 )

 as_percent()
}
