% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verbatims_fun.R
\name{verbatims_fun_single}
\alias{verbatims_fun_single}
\title{Open Ended Verbatims Function {verbatims_fun}}
\usage{
verbatims_fun_single(df, freq_var)
}
\arguments{
\item{df}{This is the name of the dataframe}

\item{freq_var}{This is the name or names of the open ended variables}
}
\value{
Returns a data frame with the variable name, variable label, and individual responses. If there is no label it just says "no label"
}
\description{
This function takes open ended variables and compiles them to be used in topline reports.
}
